blogs can now be deleted using a valid token

+++ part4/bloglist/controllers/blogs.js
@@ -40,8 +40,28 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request,response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if(!user) {
+    return response.status(401).json({ error: 'UserId missing or not valid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const userid = decodedToken.id
+
+  if(blog.user.toString() === userid.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'token invalid'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
